Consolidate shared infrastructure and fix lifecycle races
Unify file and content operations, URI grant ownership, typed integration requests, and bounded process handling. Remove obsolete paths and keep framework observations shared across UI and automation.

Fix window launch and cleanup ownership, stale UI callbacks, terminal backpressure and shutdown, native input resynchronization, and diagnostic lifetime handling. Isolate JDK test support and kernel shared sources, and add native fixtures to CI.

Verify with 1718 JVM tests, 21 native scenarios, APK boundary checks, and phone, simulated, and wired self-tests with zero failures. Document the pre-existing simulated Close transition-performance issue separately.

+++ app/src/main/java/io/github/mekhontsev/magicdesk/DesktopPlacementEngine.java
@@ -73,7 +73,7 @@ final class DesktopPlacementEngine {
             final int preferredColumn,
             final int preferredRow) {
         DesktopPlacement best = null;
-        int bestDistance = Integer.MAX_VALUE;
+        long bestDistance = Long.MAX_VALUE;
         final int safeColumnSpan = Math.max(1, Math.min(columns, columnSpan));
         final int safeRowSpan = Math.max(1, Math.min(rows, rowSpan));
         for (int row = 0; row <= rows - safeRowSpan; row++) {
@@ -83,8 +83,8 @@ final class DesktopPlacementEngine {
                 if (!isAvailable(candidate, occupied, columns, rows)) {
                     continue;
                 }
-                final int distance = Math.abs(column - preferredColumn)
-                        + Math.abs(row - preferredRow);
+                final long distance = Math.abs((long) column - preferredColumn)
+                        + Math.abs((long) row - preferredRow);
                 if (best == null || distance < bestDistance) {
                     best = candidate;
                     bestDistance = distance;
